feat: complete initial Animator component
The Animator is a canvas-based animation component, which picks up on Animations exposed by AoC solutions and allows them to be viewed and explored.

The basis of an animation is a series of DrawCall[], in which each element is a function that can take a canvas context and draw to it. The animator then controls the playback (calling) of these draw calls.